+++ src/core/element_container/SGenContainer.hpp
@@ -90,7 +90,7 @@ class LS2G_API SGenContainer: public OneSideContainer_PQ, public IteratorAdder<S
             const SolverBusIdVect & id_grid_to_solver,
             const RealVect & bus_vn_kv,
             real_type sn_mva,
-            bool ac)
+            bool ac) override
             {
                 set_osc_pq_res_p();
                 set_osc_pq_res_q(ac);